+++ PrestigeRentals.Server/Application/Services/Interfaces/IVehiclePhotosService.cs
@@ -16,7 +16,7 @@ namespace PrestigeRentals.Application.Services.Interfaces
         /// </summary>
         /// <param name="vehicleId">The ID of the vehicle for which photos are to be retrieved.</param>
         /// <returns>A task representing the asynchronous operation, containing a list of Base64 encoded images.</returns>
-        Task<ActionResult<List<string>>> GetVehiclePhotosAsBase64(int vehicleId);
+        Task<ActionResult<List<string>>> GetVehiclePhotosAsBase64(long vehicleId);
 
         /// <summary>
         /// Asynchronously uploads a new vehicle photo for a given vehicle.
@@ -24,6 +24,6 @@ namespace PrestigeRentals.Application.Services.Interfaces
         /// <param name="vehicleId">The ID of the vehicle for which the photo will be uploaded.</param>
         /// <param name="base64Image">The photo to be uploaded, represented as a Base64 encoded string.</param>
         /// <returns>A task representing the asynchronous operation, containing the vehicle photo details.</returns>
-        Task<ActionResult<VehiclePhotos>> UploadVehiclePhoto(int vehicleId, string base64Image);
+        Task<ActionResult<VehiclePhotos>> UploadVehiclePhoto(long vehicleId, string base64Image);
     }
 }
